+++ rse/main/scrapers/rsnl.py
@@ -22,12 +22,14 @@ bot = logging.getLogger("rse.main.scrapers.rsnl")
 class RSNLScraper(ScraperBase):
 
     name = "rsnl"
+    matchstring = "(researchsoftwarenl|rsnl)"
 
     def __init__(self, query=None, **kwargs):
         super().__init__(query)
 
     def latest(self, paginate=False, delay=0.0):
-        """The scraper should expose a function to populate self.results with
+        """
+        The scraper should expose a function to populate self.results with
         some number of latest entries. Unlike a search, a latest scraper does
         not by default paginate.
         """
@@ -35,7 +37,8 @@ class RSNLScraper(ScraperBase):
         return self.scrape(url, delay=delay)
 
     def search(self, query, paginate=True, delay=0.0):
-        """The scraper should expose a function to populate self.results with
+        """
+        The scraper should expose a function to populate self.results with
         a listing based on matching a search criteria.
         """
         # Haven't figured out a way to search
@@ -45,7 +48,8 @@ class RSNLScraper(ScraperBase):
         return self.latest(paginate, delay=delay)
 
     def scrape(self, url, paginate=False, delay=0.0):
-        """A shared function to scrape a set of repositories. Since the JoSS
+        """
+        A shared function to scrape a set of repositories. Since the JoSS
         pages for a search and the base are the same, we can use a shared
         function.
         """
@@ -70,7 +74,8 @@ class RSNLScraper(ScraperBase):
         return self.results
 
     def create(self, database=None, config_file=None):
-        """After a scrape (whether we obtain latest or a search query) we
+        """
+        After a scrape (whether we obtain latest or a search query) we
         run create to create software repositories based on results.
         """
         from rse.main import Encyclopedia